* Cambios en los DTOs. * Implementado los test unitarios correspondientes.
