--- src/main/java/com/ahsoka/SALC/user_model/dtos/PasswordRequest.java
@@ -1,19 +1,0 @@
-package com.ahsoka.SALC.user_model.dtos;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.NotBlank;
-
-@Data
-@AllArgsConstructor
-public class PasswordRequest {
-
-    @NotNull
-    @NotBlank
-    private String email;
-
-    @NotNull
-    @NotBlank
-    private String password;
-}
