+++ src/main/java/com/ahsoka/SALC/user_model/dtos/UpdateUserRequest.java
@@ -4,24 +4,21 @@ import com.ahsoka.SALC.user_model.persistance.entity.Role;
 import com.ahsoka.SALC.user_model.persistance.entity.User;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.NotBlank;
 
 @Data
 @AllArgsConstructor
-public class NewUserRequest {
+public class UpdateUserRequest {
 
-    @NotNull
-    @NotBlank(message = "Email is mandatory!")
     private String email;
-
-    @NotNull
+    private String password;
     private Role role;
 
     public User toUser(){
         User user = new User();
         user.setEmail(email);
         user.setRole(role);
+        user.setPassword(password);
+
         return user;
     }
 }
